pyproject.toml
[project]
name = "heat-pump-calc-api"
version = "0.3.0"
description = "Heat Pump Calculator API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "numpy-financial>=1.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "simplejson>=3.20.2",
]

[dependency-groups]
dev = [
    "line-profiler>=5.0.0",
]
